+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 + 增加 Makefile
 + 增加热更新工具 `sola-hot`
 + 增加动态模块加载扩展
++ Context 获取 Store (包括 Origin 的 Store)
 
 #### Next Version